added entrypoint, fixed circ dep
