+++ langchain-core/.gitignore
@@ -74,6 +74,10 @@ messages/tool.cjs
 messages/tool.js
 messages/tool.d.ts
 messages/tool.d.cts
+messages/utils.cjs
+messages/utils.js
+messages/utils.d.ts
+messages/utils.d.cts
 output_parsers.cjs
 output_parsers.js
 output_parsers.d.ts
